refactor: extract shared RPC API methods into RpcClientMixin
- Create rpc_client_mixin.py with abstract _send_request and all 17 shared API methods plus 4 _parse_* helpers
- RpcClient now inherits RpcClientMixin, keeping only sync ZMQ transport layer
- RpcAsyncClient now inherits RpcClientMixin, keeping only async transport and *_async methods
- Eliminates ~200 lines of duplicated request-building and response-parsing logic
