+++ maru_handler/rpc_async_client.py
@@ -33,28 +33,19 @@ import zmq
 import zmq.asyncio
 
 from maru_common import (
-    AllocationManagerStats,
     BatchExistsKVResponse,
-    BatchLookupKVResponse,
-    BatchPinKVResponse,
     BatchRegisterKVResponse,
-    BatchUnpinKVResponse,
-    GetStatsResponse,
-    KVManagerStats,
-    ListAllocationsResponse,
-    LookupKVResponse,
-    LookupResult,
     MessageType,
-    RequestAllocResponse,
     Serializer,
 )
 from maru_common.protocol import HEADER_SIZE, MessageHeader
-from maru_shm import MaruHandle
+
+from .rpc_client_mixin import RpcClientMixin
 
 logger = logging.getLogger(__name__)
 
 
-class RpcAsyncClient:
+class RpcAsyncClient(RpcClientMixin):
     """
     ZeroMQ-based async RPC client using asyncio event loop.
 
@@ -297,216 +288,6 @@ class RpcAsyncClient:
         return asyncio.run_coroutine_threadsafe(
             self._send_async(msg_type, data), self._loop
         )
-
-    # =========================================================================
-    # Response Parsing Helpers
-    # =========================================================================
-
-    @staticmethod
-    def _parse_request_alloc(response: dict) -> RequestAllocResponse:
-        if not response.get("success"):
-            return RequestAllocResponse(
-                success=False, error=response.get("error", "Unknown error")
-            )
-        handle_data = response.get("handle", {})
-        handle = MaruHandle.from_dict(handle_data) if handle_data else None
-        return RequestAllocResponse(success=True, handle=handle)
-
-    @staticmethod
-    def _parse_lookup_kv(response: dict) -> LookupKVResponse:
-        found = response.get("found", False)
-        if not found:
-            return LookupKVResponse(found=False)
-        handle = None
-        if "handle" in response:
-            handle = MaruHandle.from_dict(response["handle"])
-        return LookupKVResponse(
-            found=True,
-            handle=handle,
-            kv_offset=response.get("kv_offset", 0),
-            kv_length=response.get("kv_length", 0),
-        )
-
-    @staticmethod
-    def _parse_list_allocations(response: dict) -> ListAllocationsResponse:
-        if not response.get("success"):
-            return ListAllocationsResponse(
-                success=False, error=response.get("error", "Unknown error")
-            )
-        handles = [
-            MaruHandle.from_dict(alloc_data)
-            for alloc_data in response.get("allocations", [])
-        ]
-        return ListAllocationsResponse(success=True, allocations=handles)
-
-    @staticmethod
-    def _parse_batch_lookup_kv(response: dict) -> BatchLookupKVResponse:
-        entries = []
-        for entry_data in response.get("entries", []):
-            found = entry_data.get("found", False)
-            handle = None
-            if found and "handle" in entry_data and entry_data["handle"]:
-                handle = MaruHandle.from_dict(entry_data["handle"])
-            entries.append(
-                LookupResult(
-                    found=found,
-                    handle=handle,
-                    kv_offset=entry_data.get("kv_offset", 0),
-                    kv_length=entry_data.get("kv_length", 0),
-                )
-            )
-        return BatchLookupKVResponse(entries=entries)
-
-    # =========================================================================
-    # Allocation Management (blocking)
-    # =========================================================================
-
-    def request_alloc(self, instance_id: str, size: int) -> RequestAllocResponse:
-        """Request a new memory allocation."""
-        response = self._send_request(
-            MessageType.REQUEST_ALLOC,
-            {"instance_id": instance_id, "size": size},
-        )
-        return self._parse_request_alloc(response)
-
-    def list_allocations(
-        self, exclude_instance_id: str | None = None
-    ) -> ListAllocationsResponse:
-        """List all active allocations from the server."""
-        data: dict = {}
-        if exclude_instance_id:
-            data["exclude_instance_id"] = exclude_instance_id
-        response = self._send_request(MessageType.LIST_ALLOCATIONS, data)
-        return self._parse_list_allocations(response)
-
-    def return_alloc(self, instance_id: str, region_id: int) -> bool:
-        """Return an allocation back to the server."""
-        response = self._send_request(
-            MessageType.RETURN_ALLOC,
-            {"instance_id": instance_id, "region_id": region_id},
-        )
-        return response.get("success", False)
-
-    # =========================================================================
-    # KV Operations (blocking)
-    # =========================================================================
-
-    def register_kv(
-        self, key: str, region_id: int, kv_offset: int, kv_length: int
-    ) -> bool:
-        """Register a KV entry."""
-        response = self._send_request(
-            MessageType.REGISTER_KV,
-            {
-                "key": key,
-                "region_id": region_id,
-                "kv_offset": kv_offset,
-                "kv_length": kv_length,
-            },
-        )
-        if "error" in response:
-            raise ConnectionError(f"register_kv RPC failed: {response['error']}")
-        return response.get("is_new", False)
-
-    def lookup_kv(self, key: str) -> LookupKVResponse:
-        """Lookup a KV entry by key."""
-        response = self._send_request(MessageType.LOOKUP_KV, {"key": key})
-        return self._parse_lookup_kv(response)
-
-    def exists_kv(self, key: str) -> bool:
-        """Check if a KV entry exists."""
-        response = self._send_request(MessageType.EXISTS_KV, {"key": key})
-        return response.get("exists", False)
-
-    def pin_kv(self, key: str) -> bool:
-        """Check if a KV entry exists and pin it atomically."""
-        response = self._send_request(MessageType.PIN_KV, {"key": key})
-        return response.get("exists", False)
-
-    def unpin(self, key: str) -> bool:
-        """Unpin a KV entry, making it eligible for eviction."""
-        response = self._send_request(MessageType.UNPIN_KV, {"key": key})
-        return response.get("success", False)
-
-    def delete_kv(self, key: str) -> bool:
-        """Delete a KV entry."""
-        response = self._send_request(MessageType.DELETE_KV, {"key": key})
-        return response.get("success", False)
-
-    # =========================================================================
-    # Batch KV Operations (blocking)
-    # =========================================================================
-
-    def batch_register_kv(
-        self, entries: list[tuple[str, int, int, int]]
-    ) -> BatchRegisterKVResponse:
-        """Register multiple KV entries in a single RPC call."""
-        entries_data = [
-            {
-                "key": key,
-                "region_id": region_id,
-                "kv_offset": kv_offset,
-                "kv_length": kv_length,
-            }
-            for key, region_id, kv_offset, kv_length in entries
-        ]
-        response = self._send_request(
-            MessageType.BATCH_REGISTER_KV, {"entries": entries_data}
-        )
-        return BatchRegisterKVResponse(
-            success=response.get("success", False),
-            results=response.get("results", []),
-        )
-
-    def batch_lookup_kv(self, keys: list[str]) -> BatchLookupKVResponse:
-        """Lookup multiple KV entries in a single RPC call."""
-        response = self._send_request(MessageType.BATCH_LOOKUP_KV, {"keys": keys})
-        return self._parse_batch_lookup_kv(response)
-
-    def batch_exists_kv(self, keys: list[str]) -> BatchExistsKVResponse:
-        """Check existence of multiple KV entries in a single RPC call."""
-        response = self._send_request(MessageType.BATCH_EXISTS_KV, {"keys": keys})
-        return BatchExistsKVResponse(results=response.get("results", []))
-
-    def batch_pin_kv(self, keys: list[str]) -> BatchPinKVResponse:
-        """Check existence and pin multiple KV entries in a single RPC call."""
-        response = self._send_request(MessageType.BATCH_PIN_KV, {"keys": keys})
-        return BatchPinKVResponse(results=response.get("results", []))
-
-    def batch_unpin(self, keys: list[str]) -> BatchUnpinKVResponse:
-        """Unpin multiple KV entries in a single RPC call."""
-        response = self._send_request(MessageType.BATCH_UNPIN_KV, {"keys": keys})
-        return BatchUnpinKVResponse(results=response.get("results", []))
-
-    # =========================================================================
-    # Admin Operations (blocking)
-    # =========================================================================
-
-    def get_stats(self) -> GetStatsResponse:
-        """Get server statistics."""
-        response = self._send_request(MessageType.GET_STATS, {})
-        kv_data = response.get("kv_manager", {})
-        alloc_data = response.get("allocation_manager", {})
-        return GetStatsResponse(
-            kv_manager=KVManagerStats(
-                total_entries=kv_data.get("total_entries", 0),
-                total_size=kv_data.get("total_size", 0),
-            ),
-            allocation_manager=AllocationManagerStats(
-                num_allocations=alloc_data.get("num_allocations", 0),
-                total_allocated=alloc_data.get("total_allocated", 0),
-                active_clients=alloc_data.get("active_clients", 0),
-            ),
-        )
-
-    def heartbeat(self) -> bool:
-        """Send heartbeat to server."""
-        response = self._send_request(MessageType.HEARTBEAT, {})
-        return "error" not in response
-
-    def handshake(self) -> dict:
-        """Perform handshake with server. Returns server config (rm_address, etc.)."""
-        return self._send_request(MessageType.HANDSHAKE, {})
 
     # =========================================================================
     # Non-blocking Async API (*_async methods)
